Move special case responders into own package
