+++ src/main/java/httpserver/responder/special/CoffeeResponder.java
@@ -1,7 +1,8 @@
-package httpserver.responder;
+package httpserver.responder.special;
 
 import httpserver.AppConfig;
 import httpserver.Request;
+import httpserver.responder.Responder;
 import httpserver.response.FourEighteenResponse;
 import httpserver.response.Response;
