+++ src/main/java/httpserver/responder/special/CookieResponder.java
@@ -1,8 +1,9 @@
-package httpserver.responder;
+package httpserver.responder.special;
 
 import httpserver.AppConfig;
 import httpserver.Request;
 import httpserver.header.Header;
+import httpserver.responder.Responder;
 import httpserver.response.OkResponse;
 import httpserver.response.Response;
