+++ src/main/java/httpserver/responder/special/LogsResponder.java
@@ -1,8 +1,9 @@
-package httpserver.responder;
+package httpserver.responder.special;
 
 import httpserver.AppConfig;
 import httpserver.Authorizer;
 import httpserver.Request;
+import httpserver.responder.Responder;
 import httpserver.response.OkResponse;
 import httpserver.response.Response;
 import httpserver.response.UnauthorizedResponse;
