+++ src/main/java/httpserver/responder/special/RedirectResponder.java
@@ -1,7 +1,8 @@
-package httpserver.responder;
+package httpserver.responder.special;
 
 import httpserver.AppConfig;
 import httpserver.Request;
+import httpserver.responder.Responder;
 import httpserver.response.FoundResponse;
 import httpserver.response.Response;
