+++ src/main/java/httpserver/responder/special/TeaResponder.java
@@ -1,7 +1,8 @@
-package httpserver.responder;
+package httpserver.responder.special;
 
 import httpserver.AppConfig;
 import httpserver.Request;
+import httpserver.responder.Responder;
 import httpserver.response.OkResponse;
 import httpserver.response.Response;
