+++ src/test/java/httpserver/responder/GetResponderTest.java
@@ -6,8 +6,8 @@ import httpserver.file.Html;
 import httpserver.file.PathExaminer;
 import httpserver.header.Header;
 import httpserver.Request;
+import httpserver.header.RangeHeaderValueParser;
 import httpserver.response.Response;
-import org.junit.Ignore;
 import org.junit.Test;
 
 import java.io.IOException;
